+++ src/WinPure/Views/StartupView.xaml.cs
@@ -1,0 +1,8 @@
+using System.Windows.Controls;
+
+namespace WinPure.Views;
+
+public partial class StartupView : UserControl
+{
+    public StartupView() => InitializeComponent();
+}